+++ libtesting-thirdparty/src/main/java/com/mapbox/navigation/testing/factories/DirectionsResponseFactories.kt
@@ -147,6 +147,7 @@ fun createRouteLegAnnotation(
     duration: List<Double> = listOf(2.0, 2.0),
     maxSpeed: List<MaxSpeed> = listOf(createMaxSpeed(40), createMaxSpeed(60)),
     speed: List<Double> = listOf(40.4, 60.7),
+    stateOfCharge: List<Int> = listOf(80, 79)
 ): LegAnnotation {
     return LegAnnotation.builder()
         .distance(distance)
@@ -155,6 +156,15 @@ fun createRouteLegAnnotation(
         .congestionNumeric(congestionNumeric)
         .maxspeed(maxSpeed)
         .speed(speed)
+        .unrecognizedJsonProperties(
+            mapOf(
+                "state_of_charge" to JsonArray().apply {
+                    stateOfCharge.forEach {
+                        add(JsonPrimitive(it))
+                    }
+                }
+            )
+        )
         .build()
 }